+++ src/frameName.cpp
@@ -19,7 +19,84 @@
 #include <stdlib.h>
 #include <string.h>
 #include "frameName.h"
+#include "vmStructs.h"
 
+
+FrameName::FrameName(bool simple, bool dotted, bool use_thread_names) :
+    _cache(),
+    _simple(simple),
+    _dotted(dotted),
+    _thread_count(0),
+    _threads(NULL) {
+
+    // Require printf to use standard C format regardless of system locale
+    _saved_locale = uselocale(newlocale(LC_NUMERIC_MASK, "C", (locale_t)0));
+
+    if (use_thread_names && VMThread::available()) {
+        initThreadMap();
+    }
+}
+
+FrameName::~FrameName() {
+    jvmtiEnv* jvmti = VM::jvmti();
+    for (int i = 0; i < _thread_count; i++) {
+        jvmti->Deallocate((unsigned char*)_threads[i]._name);
+    }
+    free(_threads);
+
+    freelocale(uselocale(_saved_locale));
+}
+
+void FrameName::initThreadMap() {
+    JNIEnv* env = VM::jni();
+    jclass threadClass = env->FindClass("java/lang/Thread");
+    if (threadClass == NULL) {
+        return;
+    }
+    jfieldID eetop = env->GetFieldID(threadClass, "eetop", "J");
+    if (eetop == NULL) {
+        return;
+    }
+
+    jvmtiEnv* jvmti = VM::jvmti();
+    jthread* thread_objects;
+    if (jvmti->GetAllThreads(&_thread_count, &thread_objects) != 0) {
+        return;
+    }
+
+    _threads = (ThreadId*)calloc(_thread_count, sizeof(ThreadId));
+
+    // Create a map [OS thread ID] -> [Java thread name] backed by a sorted array
+    for (int i = 0; i < _thread_count; i++) {
+        VMThread* vm_thread = (VMThread*)(uintptr_t)env->GetLongField(thread_objects[i], eetop);
+        jvmtiThreadInfo thread_info;
+        if (vm_thread != NULL && jvmti->GetThreadInfo(thread_objects[i], &thread_info) == 0) {
+            _threads[i]._id = vm_thread->osThreadId();
+            _threads[i]._name = thread_info.name;
+        }
+    }
+
+    qsort(_threads, _thread_count, sizeof(ThreadId), ThreadId::comparator);
+    jvmti->Deallocate((unsigned char*)thread_objects);
+}
+
+const char* FrameName::findThreadName(int tid) {
+    int low = 0;
+    int high = _thread_count - 1;
+
+    while (low <= high) {
+        int mid = (unsigned int)(low + high) >> 1;
+        if (_threads[mid]._id < tid) {
+            low = mid + 1;
+        } else if (_threads[mid]._id > tid) {
+            high = mid - 1;
+        } else {
+            return _threads[mid]._name;
+        }
+    }
+
+    return NULL;
+}
 
 const char* FrameName::cppDemangle(const char* name) {
     if (name != NULL && name[0] == '_' && name[1] == 'Z') {
@@ -34,12 +111,36 @@ const char* FrameName::cppDemangle(const char* name) {
     return name;
 }
 
-char* FrameName::javaClassName(VMKlass* klass) {
-    VMSymbol* symbol = klass->name();
-    return javaClassName(symbol->body(), symbol->length(), true);
+char* FrameName::javaMethodName(jmethodID method) {
+    jclass method_class;
+    char* class_name = NULL;
+    char* method_name = NULL;
+    char* result;
+
+    jvmtiEnv* jvmti = VM::jvmti();
+    jvmtiError err;
+
+    if ((err = jvmti->GetMethodName(method, &method_name, NULL, NULL)) == 0 &&
+        (err = jvmti->GetMethodDeclaringClass(method, &method_class)) == 0 &&
+        (err = jvmti->GetClassSignature(method_class, &class_name, NULL)) == 0) {
+        // Trim 'L' and ';' off the class descriptor like 'Ljava/lang/Object;'
+        result = javaClassName(class_name + 1, strlen(class_name) - 2, _simple, _dotted);
+        strcat(result, ".");
+        strcat(result, method_name);
+    } else {
+        snprintf(_buf, sizeof(_buf), "[jvmtiError %d]", err);
+        result = _buf;
+    }
+
+    jvmti->Deallocate((unsigned char*)class_name);
+    jvmti->Deallocate((unsigned char*)method_name);
+
+    return result;
 }
 
-char* FrameName::javaClassName(const char* symbol, int length, bool dotted) {
+char* FrameName::javaClassName(const char* symbol, int length, bool simple, bool dotted) {
+    char* result = _buf;
+
     int array_dimension = 0;
     while (*symbol == '[') {
         array_dimension++;
@@ -47,75 +148,85 @@ char* FrameName::javaClassName(const char* symbol, int length, bool dotted) {
     }
 
     if (array_dimension == 0) {
-        strncpy(_buf, symbol, length);
-        _buf[length] = 0;
+        strncpy(result, symbol, length);
+        result[length] = 0;
     } else {
         switch (*symbol) {
-            case 'B': strcpy(_buf, "byte");    break;
-            case 'C': strcpy(_buf, "char");    break;
-            case 'I': strcpy(_buf, "int");     break;
-            case 'J': strcpy(_buf, "long");    break;
-            case 'S': strcpy(_buf, "short");   break;
-            case 'Z': strcpy(_buf, "boolean"); break;
-            case 'F': strcpy(_buf, "float");   break;
-            case 'D': strcpy(_buf, "double");  break;
+            case 'B': strcpy(result, "byte");    break;
+            case 'C': strcpy(result, "char");    break;
+            case 'I': strcpy(result, "int");     break;
+            case 'J': strcpy(result, "long");    break;
+            case 'S': strcpy(result, "short");   break;
+            case 'Z': strcpy(result, "boolean"); break;
+            case 'F': strcpy(result, "float");   break;
+            case 'D': strcpy(result, "double");  break;
             default:
                 length -= array_dimension + 2;
-                strncpy(_buf, symbol + 1, length);
-                _buf[length] = 0;
+                strncpy(result, symbol + 1, length);
+                result[length] = 0;
         }
 
         do {
-            strcat(_buf, "[]");
+            strcat(result, "[]");
         } while (--array_dimension > 0);
     }
 
+    if (simple) {
+        for (char* s = result; *s; s++) {
+            if (*s == '/') result = s + 1;
+        }
+    }
+
     if (dotted) {
-        for (char* s = _buf; *s; s++) {
+        for (char* s = result; *s; s++) {
             if (*s == '/') *s = '.';
         }
     }
 
-    return _buf;
+    return result;
 }
 
-FrameName::FrameName(ASGCT_CallFrame& frame, bool dotted) {
+const char* FrameName::name(ASGCT_CallFrame& frame) {
     if (frame.method_id == NULL) {
-        _str = "[unknown]";
+        return "[unknown]";
+    }
 
-    } else if (frame.bci == BCI_NATIVE_FRAME) {
-        _str = cppDemangle((const char*)frame.method_id);
+    switch (frame.bci) {
+        case BCI_NATIVE_FRAME:
+            return cppDemangle((const char*)frame.method_id);
 
-    } else if (frame.bci == BCI_KLASS) {
-        VMKlass* alloc_class = (VMKlass*)frame.method_id;
-        _str = strcat(javaClassName(alloc_class), dotted ? "" : "_[i]");
-
-    } else if (frame.bci == BCI_KLASS_OUTSIDE_TLAB) {
-        VMKlass* alloc_class = (VMKlass*)((uintptr_t)frame.method_id ^ 1);
-        _str = strcat(javaClassName(alloc_class), dotted ? " (out)" : "_[k]");
-
-    } else {
-        jclass method_class;
-        char* class_name = NULL;
-        char* method_name = NULL;
-
-        jvmtiEnv* jvmti = VM::jvmti();
-        jvmtiError err;
-
-        if ((err = jvmti->GetMethodName(frame.method_id, &method_name, NULL, NULL)) == 0 &&
-            (err = jvmti->GetMethodDeclaringClass(frame.method_id, &method_class)) == 0 &&
-            (err = jvmti->GetClassSignature(method_class, &class_name, NULL)) == 0) {
-            // Trim 'L' and ';' off the class descriptor like 'Ljava/lang/Object;'
-            char* s = javaClassName(class_name + 1, strlen(class_name) - 2, dotted);
-            strcat(s, ".");
-            strcat(s, method_name);
-            _str = s;
-        } else {
-            snprintf(_buf, sizeof(_buf), "[jvmtiError %d]", err);
-            _str = _buf;
+        case BCI_SYMBOL: {
+            VMSymbol* symbol = (VMSymbol*)frame.method_id;
+            char* class_name = javaClassName(symbol->body(), symbol->length(), _simple, true);
+            return strcat(class_name, _dotted ? "" : "_[i]");
         }
 
-        jvmti->Deallocate((unsigned char*)class_name);
-        jvmti->Deallocate((unsigned char*)method_name);
+        case BCI_SYMBOL_OUTSIDE_TLAB: {
+            VMSymbol* symbol = (VMSymbol*)((uintptr_t)frame.method_id ^ 1);
+            char* class_name = javaClassName(symbol->body(), symbol->length(), _simple, true);
+            return strcat(class_name, _dotted ? " (out)" : "_[k]");
+        }
+
+        case BCI_THREAD_ID: {
+            int tid = (int)(uintptr_t)frame.method_id;
+            const char* name = findThreadName(tid);
+            if (name != NULL) {
+                return name;
+            }
+
+            snprintf(_buf, sizeof(_buf), "[thread %d]", tid);
+            return _buf;
+        }
+
+        default: {
+            JMethodCache::iterator it = _cache.lower_bound(frame.method_id);
+            if (it != _cache.end() && it->first == frame.method_id) {
+                return it->second.c_str();
+            }
+
+            const char* newName = javaMethodName(frame.method_id);
+            _cache.insert(it, JMethodCache::value_type(frame.method_id, newName));
+            return newName;
+        }
     }
 }