+++ src/profiler.h
@@ -40,12 +40,6 @@ static inline int cmp64(u64 a, u64 b) {
 }
 
 
-union CallTraceBuffer {
-    ASGCT_CallFrame _asgct_frames[MAX_STACK_FRAMES];
-    jvmtiFrameInfo _jvmti_frames[MAX_STACK_FRAMES];
-};
-
-
 class CallTraceSample {
   private:
     u64 _samples;
@@ -130,11 +124,12 @@ class Profiler {
     MethodSample _methods[MAX_CALLTRACES];
 
     SpinLock _locks[CONCURRENCY_LEVEL];
-    CallTraceBuffer _calltrace_buffer[CONCURRENCY_LEVEL];
+    ASGCT_CallFrame _calltrace_buffer[CONCURRENCY_LEVEL][MAX_STACK_FRAMES];
     ASGCT_CallFrame* _frame_buffer;
     int _frame_buffer_size;
     volatile int _frame_buffer_index;
     bool _frame_buffer_overflow;
+    bool _threads;
 
     SpinLock _jit_lock;
     const void* _jit_min_address;
@@ -150,19 +145,19 @@ class Profiler {
     void updateJitRange(const void* min_address, const void* max_address);
 
     const char* findNativeMethod(const void* address);
-    int getNativeTrace(void* ucontext, ASGCT_CallFrame* frames);
+    int getNativeTrace(int tid, ASGCT_CallFrame* frames);
     int getJavaTraceAsync(void* ucontext, ASGCT_CallFrame* frames, int max_depth);
-    int getJavaTraceJVMTI(jvmtiFrameInfo* jvmti_frames, ASGCT_CallFrame* frames, int max_depth);
     int makeEventFrame(ASGCT_CallFrame* frames, jint event_type, jmethodID event);
     bool fillTopFrame(const void* pc, ASGCT_CallFrame* frame);
+    bool addressInCode(const void* pc);
     u64 hashCallTrace(int num_frames, ASGCT_CallFrame* frames);
     void storeCallTrace(int num_frames, ASGCT_CallFrame* frames, u64 counter);
     void copyToFrameBuffer(int num_frames, ASGCT_CallFrame* frames, CallTraceSample* trace);
     u64 hashMethod(jmethodID method);
     void storeMethod(jmethodID method, jint bci, u64 counter);
+    void initStateLock();
     void resetSymbols();
     void setSignalHandler();
-    void runInternal(Arguments& args, std::ostream& out);
 
   public:
     static Profiler _instance;
@@ -176,7 +171,7 @@ class Profiler {
         _java_methods(),
         _runtime_stubs("[stubs]"),
         _native_lib_count(0) {
-        pthread_mutex_init(&_state_lock, NULL);
+        initStateLock();
     }
 
     u64 total_samples() { return _total_samples; }
@@ -184,19 +179,17 @@ class Profiler {
     time_t uptime()     { return time(NULL) - _start_time; }
 
     void run(Arguments& args);
-    Error start(const char* event, long interval, int frame_buffer_size);
+    void runInternal(Arguments& args, std::ostream& out);
+    void shutdown(Arguments& args);
+    Error start(const char* event, long interval, int frame_buffer_size, bool threads);
     Error stop();
     void dumpSummary(std::ostream& out);
-    void dumpCollapsed(std::ostream& out, Counter counter);
+    void dumpCollapsed(std::ostream& out, Arguments& args);
+    void dumpFlameGraph(std::ostream& out, Arguments& args);
     void dumpTraces(std::ostream& out, int max_traces);
     void dumpFlat(std::ostream& out, int max_methods);
     void recordSample(void* ucontext, u64 counter, jint event_type, jmethodID event);
     NativeCodeCache* jvmLibrary();
-
-    static void JNICALL VMDeath(jvmtiEnv* jvmti, JNIEnv* jni) {
-        MutexLocker ml(_instance._state_lock);
-        _instance._state = TERMINATED;
-    }
 
     // CompiledMethodLoad is also needed to enable DebugNonSafepoints info by default
     static void JNICALL CompiledMethodLoad(jvmtiEnv* jvmti, jmethodID method,