+++ src/vmEntry.h
@@ -22,9 +22,10 @@
 
 // Denotes ASGCT_CallFrame where method_id has special meaning (not jmethodID)
 enum ASGCT_CallFrameType {
-    BCI_NATIVE_FRAME       = -10,  // method_id is native function name (char*)
-    BCI_KLASS              = -11,  // method_id is Klass descriptor (VMKlass*)
-    BCI_KLASS_OUTSIDE_TLAB = -12,  // VMKlass* specifically for allocations outside TLAB
+    BCI_NATIVE_FRAME        = -10,  // method_id is native function name (char*)
+    BCI_SYMBOL              = -11,  // method_id is VMSymbol*
+    BCI_SYMBOL_OUTSIDE_TLAB = -12,  // VMSymbol* specifically for allocations outside TLAB
+    BCI_THREAD_ID           = -13,  // method_id designates a thread
 };
 
 typedef struct {
@@ -63,9 +64,8 @@ class VM {
         return _vm->GetEnv((void**)&jni, JNI_VERSION_1_6) == 0 ? jni : NULL;
     }
 
-    static void JNICALL VMInit(jvmtiEnv* jvmti, JNIEnv* jni, jthread thread) {
-        loadAllMethodIDs(jvmti);
-    }
+    static void JNICALL VMInit(jvmtiEnv* jvmti, JNIEnv* jni, jthread thread);
+    static void JNICALL VMDeath(jvmtiEnv* jvmti, JNIEnv* jni);
 
     static void JNICALL ClassLoad(jvmtiEnv* jvmti, JNIEnv* jni, jthread thread, jclass klass) {
         // Needed only for AsyncGetCallTrace support