+++ src/vmStructs.h
@@ -17,6 +17,7 @@
 #ifndef _VMSTRUCTS_H
 #define _VMSTRUCTS_H
 
+#include <stdint.h>
 #include "codeCache.h"
 
 
@@ -26,6 +27,9 @@ class VMStructs {
     static int _symbol_length_offset;
     static int _symbol_body_offset;
     static int _class_klass_offset;
+    static int _thread_osthread_offset;
+    static int _osthread_id_offset;
+    static bool _has_perm_gen;
 
     const char* at(int offset) {
         return (const char*)this + offset;
@@ -39,6 +43,10 @@ class VMStructs {
             && _symbol_length_offset >= 0
             && _symbol_body_offset >= 0
             && _class_klass_offset >= 0;
+    }
+
+    static bool hasPermGen() {
+        return _has_perm_gen;
     }
 };
 
@@ -56,6 +64,15 @@ class VMSymbol : VMStructs {
 
 class VMKlass : VMStructs {
   public:
+    static VMKlass* fromHandle(uintptr_t handle) {
+        if (_has_perm_gen) {
+            // On JDK 7 KlassHandle is a pointer to klassOop, hence one more indirection
+            return (VMKlass*)(*(uintptr_t**)handle + 2);
+        } else {
+            return (VMKlass*)handle;
+        }
+    }
+
     VMSymbol* name() {
         return *(VMSymbol**) at(_klass_name_offset);
     }
@@ -65,6 +82,18 @@ class java_lang_Class : VMStructs {
   public:
     VMKlass* klass() {
         return *(VMKlass**) at(_class_klass_offset);
+    }
+};
+
+class VMThread : VMStructs {
+  public:
+    static bool available() {
+        return _thread_osthread_offset >= 0 && _osthread_id_offset >= 0;
+    }
+
+    int osThreadId() {
+        const char* osthread = *(const char**) at(_thread_osthread_offset);
+        return *(int*)(osthread + _osthread_id_offset);
     }
 };