+++ CHANGELOG.md
@@ -1,13 +1,29 @@
 # Changelog
 
-## [Unreleased]
+## [1.2] - 2018-03-05
+
+### Features
+ - Produce SVG files out of the box; flamegraph.pl is no longer needed
+ - Profile ReentrantLock contention
+ - Java API
+ 
+### Improvements
+ - Allocation and Lock profiler now works on JDK 7, too
+ - Faster dumping of results
+
+### Changes
+ - `total` counter of allocation profiler now measures heap pressure (like JMC)
+
+## [1.1] - 2017-12-03
 
 ### Features
  - Linux Perf Events profiling: CPU cycles, cache misses, branch misses, page faults, context switches etc.
  - Kernel tracepoints support
  - Contended monitor (aka intrinsic lock) profiling
+ - Individual thread profiles
 
 ### Improvements
+ - Profiler can engage at JVM start and automatically dump results on exit
  - `list` command-line option to list supported events
  - Automatically find target process ID with `jps` tool
  - An option to include counter value in `collapsed` output
@@ -16,7 +32,7 @@
 
 ### Changes
  - Replaced `-m` modes with `-e` events
- - Interval changed from 'int' to 'long'
+ - Interval changed from `int` to `long`
 
 ## [1.0] - 2017-10-09